Move Phoenix.HTML.Tag.csrf_token_value to Phoenix.HTML
