+++ test/phoenix_html/tag_test.exs
@@ -182,28 +182,4 @@ defmodule Phoenix.HTML.TagTest do
     assert safe_to_string(csrf_input_tag(url, foo: "bar")) ==
              ~s(<input foo="bar" name="_csrf_token" type="hidden" value="#{csrf_token}">)
   end
-
-  describe "csrf_token_value" do
-    def custom_csrf(to, extra), do: "#{extra}:#{to}"
-
-    test "with default" do
-      assert csrf_token_value("/") == Plug.CSRFProtection.get_csrf_token()
-    end
-
-    test "with configured MFA" do
-      default_reader = Application.fetch_env!(:phoenix_html, :csrf_token_reader)
-
-      try do
-        Application.put_env(
-          :phoenix_html,
-          :csrf_token_reader,
-          {__MODULE__, :custom_csrf, ["extra"]}
-        )
-
-        assert csrf_token_value("/foo") == "extra:/foo"
-      after
-        Application.put_env(:phoenix_html, :csrf_token_reader, default_reader)
-      end
-    end
-  end
 end
